Refactored quickqueck code
